pyproject.toml
[project]
name = "memorybot"
version = "0.1.0"
description = "Discord chatbot with memory integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.4.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "openai>=1.43.0",
  "SQLAlchemy>=2.0.32",
  "aiosqlite>=0.20.0",
  "greenlet>=3.2.4",
  "tavily-python>=0.3.1",
]

[project.scripts]
memorybot = "memorybot.main:run"
bot = "memorybot.main:run"
tavily-sample = "memorybot.scripts.tavily_sample:main"

[tool.uv]
dev-dependencies = []
package = true

[tool.uv.sources]

 
